Introduce core base class that contains convention
